Build test-macros in a build script
