+++ crates/proc_macro_test/imp/.gitignore
@@ -1,0 +1,2 @@
+target/
+Cargo.lock
